+++ packages/backend/src/program/domain/preprocessProgram.ts
@@ -1,0 +1,47 @@
+import { ProgramData, QuestionnaireRoute } from '@tee/web/src/types'
+import { PublicodesInputData, PublicodesKeys, PublicodesQuestionnaireRoute, QuestionnaireData } from './types'
+
+/** preprocesses the data gathered from the questionnaire into variables
+ * needed by publicodes */
+export const preprocessInputForPublicodes = (
+  questionnaireData: QuestionnaireData,
+  programData: ProgramData,
+  currentDate: string
+): PublicodesInputData => {
+  const publicodesData: PublicodesInputData = {
+    ...questionnaireData,
+    [PublicodesKeys.CurrentDate]: currentDate
+  }
+
+  if (questionnaireData.codeNaf) publicodesData['entreprise . code NAF'] = enquotePublicodesLiteralString(questionnaireData.codeNaf)
+
+  const route = questionnaireData.questionnaire_route
+  if (route) {
+    publicodesData[PublicodesKeys.QuestionnaireRoute] = convertQuestionnaireRoute(route)
+  }
+
+  if (programData['début de validité']) {
+    publicodesData['dispositif . début de validité'] = programData['début de validité']
+  }
+  if (programData['fin de validité']) {
+    publicodesData['dispositif . fin de validité'] = programData['fin de validité']
+  }
+
+  return publicodesData
+}
+
+const convertQuestionnaireRoute = (route: QuestionnaireRoute): PublicodesQuestionnaireRoute => {
+  switch (route) {
+    case QuestionnaireRoute.NoSpecificGoal:
+      return PublicodesQuestionnaireRoute.NoSpecificGoal
+    case QuestionnaireRoute.SpecificGoal:
+      return PublicodesQuestionnaireRoute.SpecificGoal
+  }
+}
+
+/** for publicodes to interpret a value as a literal string, it expects
+ * extra pair of quotes, added by this function. Without it, it is interpreted as a reference to another rule
+ */
+const enquotePublicodesLiteralString = (value: string): string => {
+  return `"${value}"`
+}
